Выносит createTypeListTemplate из утилит в темлпейт
